fix(app): simplify withdrawn scene
